pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.79.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "reportlab>=4.4.1",
    "streamlit>=1.45.1",
]
